simplify tests
