+++ test_requirements.in
@@ -1,5 +1,5 @@
 -c requirements.txt
-pylint==2.7.2
+pylint==2.11.1
 black==20.8b1
 coverage==5.5
 mypy==0.812
